deployment/requirements.txt
pillow~=10.4.0
piexif~=1.1.3
